chinh sua gio hang

+++ src/java/Controller/DeleteProInCart.java
@@ -1,4 +1,3 @@
-
 package Controller;
 
 import DAO.ProductDAO;
@@ -8,6 +7,7 @@ import Model.Product;
 import java.io.IOException;
 import java.io.PrintWriter;
 import java.sql.SQLException;
+import java.util.ArrayList;
 import java.util.List;
 import java.util.logging.Level;
 import java.util.logging.Logger;
@@ -22,40 +22,44 @@ import javax.servlet.http.HttpSession;
 @WebServlet(name = "DeleteProInCart", urlPatterns = {"/DeleteProInCart"})
 public class DeleteProInCart extends HttpServlet {
 
- 
-
- 
     @Override
     protected void doGet(HttpServletRequest request, HttpServletResponse response)
             throws ServletException, IOException {
-        
+
         HttpSession session = request.getSession();
-        Customer cus =(Customer) session.getAttribute("cus");
-        
+        Customer cus = (Customer) session.getAttribute("cus");
+        List<ProCart> listPro = (ArrayList<ProCart>) session.getAttribute("prosInCart");
         try {
-            
-            String idPro =  request.getParameter("id");
-            ProductDAO proDAO = new ProductDAO();
-           // System.out.println("idPro="+idPro);
-          //  System.out.println("idCus="+cus.getId());
-            proDAO.deleteProInCart(cus.getId(),idPro);
-            
-            List<ProCart> list = proDAO.getProCartByIdCus(cus.getId());
-            request.setAttribute("cartPro", list);
-            
-            RequestDispatcher rd = request.getRequestDispatcher("cart.jsp");
-            rd.forward(request, response);
-            
+            String idPro = request.getParameter("id");
+            for (ProCart p : listPro) {
+                if (p.getId().equals(idPro)) {
+                    listPro.remove(p);
+                    break;
+                }
+            }
+            if (cus != null) {
+                ProductDAO proDAO = new ProductDAO();
+                // System.out.println("idPro="+idPro);
+                //  System.out.println("idCus="+cus.getId());
+                proDAO.deleteProInCart(cus.getId(), idPro);
+                List<ProCart> list = proDAO.getProCartByIdCus(cus.getId());
+                if (list.size() == 0) {
+                    RequestDispatcher rd = request.getRequestDispatcher("emptyCart.jsp");
+                    rd.forward(request, response);
+                }
+            }
+            session.removeAttribute("prosInCart");
+            session.setAttribute("prosInCart", listPro);
+            response.sendRedirect(request.getContextPath() + "/Cart");
+
         } catch (ClassNotFoundException ex) {
             Logger.getLogger(DeleteProInCart.class.getName()).log(Level.SEVERE, null, ex);
         } catch (SQLException ex) {
             Logger.getLogger(DeleteProInCart.class.getName()).log(Level.SEVERE, null, ex);
         }
-        
-        
+
     }
 
- 
     @Override
     protected void doPost(HttpServletRequest request, HttpServletResponse response)
             throws ServletException, IOException {
